requirements.txt
django>=2.1
django-allauth
requests
requests-oauthlib
Pillow
mysqlclient